requirements.txt
streamlit>=1.39
openai>=1.51
python-dotenv>=1.0
